requirements.txt
beautifulsoup4==4.10.0
cxxfilt==0.3.0
lxml==4.9.1
matplotlib==3.3.4
PyYAML==6.0
soupsieve==2.2.1
flake8
pep8
mypy
psutil
pytest
sphinx
sphinx_rtd_theme
configparser
coverage
atheris
